Separate generator and parser into separate modules
